+++ src/main/java/domain/piece/move/CastlePieceMove.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain.piece.move;
 
 import domain.Direction;
 import domain.JanggiBoard;
